New endpoint to allow setting arbitrary custom attributes
Implement `PATCH
:reposerver/api/v1/user_repo/proprietary-json/:target-name` accepting
a json body.

The existing target custom metadata will be merged with the provided
json body, but not deep merged.

For example, for a target with the following custom metadata:

```
{
  "name": "pkgname",
  "version": "pkgversion",
  "hardwareIds": [],
  "targetFormat": "BINARY",
  "uri": null,
  "createdAt": "2021-06-18T10:42:14Z",
  "updatedAt": "2021-06-18T10:42:14Z",
  "myobj1": {
    "nested1": "somestrvalue"
  },
  "myobj2": {
    "nested2": "somestrvalue"
  }
}

```

And a request: `PATCH
:reposerver/api/v1/user_repo/proprietary-json/:target-name` with body:

```
{
  "myobj1": {
    "nested1-2": "somevalue"
  }
}

```

The resulting custom metadata is:

```
{
  "name": "pkgname",
  "version": "pkgversion",
  "hardwareIds": [],
  "targetFormat": "BINARY",
  "uri": null,
  "createdAt": "2021-06-18T10:42:14Z",
  "updatedAt": "2021-06-18T10:42:14Z",
  "myobj1": {
    "nested1-2": "somevalue"
  },
  "myobj2": {
    "nested2": "somestrvalue"
  }
}

```

As a special case, if the json body is not json object or is an empty
object (`{}`), that object will overwrite the existing custom
attributes, except the attributes that cannot be overwritten (name,
version, uri, etc)