by default, create keys immediately when creating a repository
Previously keys were created asynchronously and the client would need
to retry until the keys were created.

Currently we don't process many of these requests. In addition, we
already use ED25519 keys by default, which are faster to generate and
therefore can easily be generated synchronously.

This means we no longer need to run the daemon to generate the keys in
the background and the tuf-keyserver deployment can be simplified.