When generating keys without a daemon, also generate root.json
This was previously not noticed because something would `GET
root.json` to make sure the keys were generated, trigerring the
generation of `root.json` implicitly. Now that we generate the keys
when the user `POSTs`, we need to generate `root.json` immediately so
other features that assume `root.json` was created work properly.