[OTA-1473] Prevent upload of new delegated targets failing pattern match
The Trusted Delegation references in a user's targets.json
have a field named "paths" which supplies a list of patterns used
by the end client(aktualizr) to determine which delegated
metadata to download via pattern matching of the target-filename.

Reposerver will now fail to allow the creation/update of delegated metadata
if any of the targets comprised in the signed metadata have target names
that don't match the specified patterns in the user's trusted-delegations
"paths" list.

Signed-off-by: Ben Clouser <ben.clouser@toradex.com>