[keyserver] accept expire-not-before parameter when fetching a root
If the current root expires before `expire-not-before`, set the
expiration date of the root and signed it again before saving it and
returning it. This requires the root keys to be online.

Previously existing behavior was kept, the root is still refreshed as
needed.

[reposerver] when setting a `expire-not-before` date via the
`expires/not-before` API, fetch the root from keyserver using the
given Instant. This forces the root to be refreshed with the new
expiration, if needed.

The request will fail if the root cannot be refreshed, for example, if
the root keys are not online, but the setting will be saved and used
by reposerver. The user is informed of this using a 502 http status
code and descriptive error message, json formatted.