Change order of package deletion to edit targets role first
Small effort to improve failures due to race conditions when editing targets metadata.

The real fix would be to rework `deleteTargetItem()` so that it only removes the target item from db if
modifying the targetsRole (and resigning it) is successful. This is MUCH MORE work!
Likely requiring a daemon process to attempt retries.

Signed-off-by: Ben Clouser <dev@benclouser.com>