Extend expiresAt when generating root with `expires-not-before`
When generating a new root.json due to the expiresAt being before the
supplied `expireNotBefore`, renew root with an expiration date of
`expiresNotBefore + defaultRoleExpire`. `defaultRoleExpire` is
currently set to 365 days.

This is to avoid users generating too many root.json versions by
GETing a root.json with `expires-not-before` set. Without this change,
a new root.json would be generated every time, each time with an
expire date set to the value of `expires-not-before`. This way, a
root.json is generated with the expire date of `expires-not-before +
defaultRoleExpire`, and subsequent requests will not generate another
root.json unless `expires-not-before` is set to a value later than
`expires-not-before + defaultNotExpire`, which by default happens only
in a year. Using the defaults, only one root.json will be generated
per year, unless the user triggers the generation explicitly.